refactor(content-details): add padding bottom ; fix flag overlaping page
